Split content action buttons into multiple files
